BellBruno

Documentation for BellBruno.

BellBruno.bell_coeff

BellBruno.bell_poly

BellBruno.note_bell_poly

BellBruno.reduce_bell_poly

BellBruno.bell_coeff — Method

bell_coeff(bp)

Computes and returns the coefficients of all Bell polynomials saved in bp.

BellBruno.bell_poly — Method

bell_poly(N::Int64)

Computes all Bell polynomials up to order of N.

Returns a Vector of N Bell polynomials (saved as vectors and matrices).

BellBruno.note_bell_poly — Method

note_bell_poly(bp, index)

Returns the Bell polynomial of index as a string.

BellBruno.reduce_bell_poly — Method

reduce_bell_poly(bp; max_order = 2 :: Int64)

Removes all monomials from a Bell polynomial with an higher order than max_order.

Returns the reduced vector of Bell polynomials.